Move DeepseekV4FP8Config
