+++ vllm/models/deepseek_v4/__init__.py
@@ -11,17 +11,16 @@ from typing import TYPE_CHECKING
 
 from vllm.platforms import current_platform
 
+from .quant_config import DeepseekV4FP8Config
+
 # Pick the per-platform implementation. The NVIDIA branch is the static
 # default that mypy sees; the ROCm branch overrides it at runtime and is
 # kept type-compatible via ``# type: ignore[assignment]``.
 if TYPE_CHECKING or not current_platform.is_rocm():
-    from .nvidia.deepseek_v4 import DeepseekV4ForCausalLM, DeepseekV4FP8Config
+    from .nvidia.deepseek_v4 import DeepseekV4ForCausalLM
     from .nvidia.deepseek_v4_mtp import DeepSeekV4MTP
 else:
-    from .amd.deepseek_v4 import (  # type: ignore[assignment]
-        DeepseekV4ForCausalLM,
-        DeepseekV4FP8Config,
-    )
+    from .amd.deepseek_v4 import DeepseekV4ForCausalLM  # type: ignore[assignment]
     from .amd.deepseek_v4_mtp import DeepSeekV4MTP  # type: ignore[assignment]
 
 __all__ = [
